requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==3.0.2
ollama==0.4.7
packaging==24.2
parsel==1.10.0
pluggy==1.5.0
pgvector==0.3.5
Protego==0.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
pytest==8.3.4
python-dotenv==1.0.1
queuelib==1.7.0
regex==2025.11.3
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
scrapy-splash==0.11.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tiktoken==0.12.0
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
w3lib==2.2.1
Werkzeug==3.1.3
zope.interface==7.2
openai>=1.0.0